pyproject.toml
[project]
name = "django-todo"
version = "0.1.0"
description = ""
authors = [
    {name = "kimjinsu",email = "creeeeeeeeeeeeeping@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "django-summernote (>=0.8.20.0,<0.9.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
